requirements.txt
flask==3.0.3
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
playwright==1.43.0
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
gunicorn==22.0.0
pytest==8.1.1
pytest-mock==3.14.0
respx==0.21.1
pytest-asyncio==0.23.6
